Make sure `esm` usage works as expected
It just doesn't work, as mocha is using ESM build and ... something's
not right or I don't know and I don't have time right now to figure this
out.

Without `esm` and on both 8.x and 9.x `mocha` versions, this happened:

1. When run, not in a watch mode, with `"type": "module"` setting
   in package.json, it worked (simply: `mocha` as script). But it
   also required the imported files to have file extensions in them.
2. When run in watch mode with settings like above ("module" and file
   extension), it did not.

Only after downgrading to 8.x and going with `esm` it all works as
expected, and not file extension needed (for the cases so far).